fixing register
