+++ frontend/src/old/oldLogin.js
@@ -1,0 +1,60 @@
+import React, { useState } from "react";
+import Form from "react-bootstrap/Form";
+import Button from "react-bootstrap/Button";
+import axios from 'axios';
+import "./Login.css";
+
+function Login(props) {
+  const email = useFormInput('');
+  const password = useFormInput('');
+  const [error, setError] = useState(null);
+  const [loading, setLoading] = useState(false);
+ 
+  // handle button click of login form
+  const handleLogin = () => {
+    const data = {
+      email: email.value, 
+      password: password.value, 
+    }
+
+    axios.post('/api/user', data)
+      .then(function (response) {
+        console.log(response);
+        console.log(response.data)
+      })
+      .catch(function (error) {
+        console.log(error);
+      });
+  }
+ 
+  return (
+    <div>
+      <h1>Login</h1>
+      <br /><br />
+      <div>
+        Email<br />
+        <input type="text" {...email} autoComplete="new-password" />
+      </div>
+      <div style={{ marginTop: 10 }}>
+        Password<br />
+        <input type="password" {...password} autoComplete="new-password" />
+      </div>
+      {error && <><small style={{ color: 'red' }}>{error}</small><br /></>}<br />
+      <input type="button" value={loading ? 'Loading...' : 'Login'} onClick={handleLogin} disabled={loading} /><br />
+    </div>
+  );
+}
+ 
+const useFormInput = initialValue => {
+  const [value, setValue] = useState(initialValue);
+ 
+  const handleChange = e => {
+    setValue(e.target.value);
+  }
+  return {
+    value,
+    onChange: handleChange
+  }
+}
+ 
+export default Login;
